+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Feat
 
 - 增加报错信息以显式提醒`POST/PATCH/PUT`方法缺少必要参数；
+- (2024-04-17)修改logger, 移除Context.Logger()方法, 但是可以通过`fastapi.Info`获得;
 
 ## 0.2.5 - (2024-03-24)